requirements.txt
opencv-python>=4.8.0
pytesseract>=0.3.10
PyMuPDF>=1.23.0
numpy>=1.24.0
setuptools>=65.0.0
Pillow>=10.0.0
pdf2image
fpdf
